Rename FooCompleter js files to match FooCommandProvider file names
