+++ packages/phoenix/src/puter-shell/completers/CommandCompleter.js
