+++ packages/phoenix/src/puter-shell/completers/OptionCompleter.js
